Strip newlines/spaces from our input program
Oftentimes brainfuck programs are formatted in a "readable" fashion
with newlines, and spaces.  If we have an input-program that
looks like this:

> ---------------------------------
> ---------------------------------
> +++----

We'd parse that naively as:

>  {Type: "-", Repeat: X}
>  {Type: "-", Repeat: X}
>  {Type: "+", Repeat: X}
>  {Type: "-", Repeat: X}

i.e. The first two lines would count as two tokens.  If we remove
the newlines before we process the input we instead get:

>  {Type: "-", Repeat: X}
>  {Type: "+", Repeat: X}
>  {Type: "-", Repeat: X}